backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
tensorflow>=2.15.0
pillow>=10.0.0
numpy>=1.24.0
